Move `Tabs` to core as `RowTabs`
This move will allow the browser extension to render tabs for a row. Also found that only two callbacks are needed to remove TabContext and PayloadContext, so did that.
