+++ packages/core/package.json
@@ -45,5 +45,8 @@
     }
   },
   "types": "./dist/src/main.d.ts",
-  "style": "./dist/style.css"
+  "style": "./dist/style.css",
+  "dependencies": {
+    "react-json-tree": "^0.18.0"
+  }
 }
