+++ packages/core/src/main.ts
@@ -2,6 +2,7 @@ import { Meter } from "./Meter";
 import { ClientReferenceRow } from "./rows/ClientReferenceRow";
 import { HintRow } from "./rows/HintRow";
 import { TreeRow } from "./rows/TreeRow";
+import { RowTabs } from "./tabs/row/RowTabs";
 import { parse, refineRowType, lexer, splitToCleanRows } from "./parse";
 
 export {
@@ -13,4 +14,5 @@ export {
   refineRowType,
   lexer,
   splitToCleanRows,
+  RowTabs,
 };
