+++ packages/core/src/tabs/row/stringToKiloBytes.ts
