+++ packages/website/app/Parser.tsx
@@ -1,31 +1,9 @@
 "use client";
 
-import React, {
-  ChangeEvent,
-  ReactNode,
-  useContext,
-  useEffect,
-  useState,
-  useTransition,
-} from "react";
-import { JSONTree } from "react-json-tree";
-
-import * as Ariakit from "@ariakit/react";
+import React, { ChangeEvent, useEffect, useState } from "react";
 import { ErrorBoundary } from "react-error-boundary";
-import {
-  Meter,
-  TreeRow,
-  ClientReferenceRow,
-  HintRow,
-  lexer,
-  parse,
-  refineRowType,
-  splitToCleanRows,
-} from "@rsc-parser/core";
+import { RowTabs } from "@rsc-parser/core";
 import { GenericErrorBoundaryFallback } from "./GenericErrorBoundaryFallback";
-import { stringToKiloBytes } from "./stringtoKiloBytes";
-import { PayloadContext } from "./PayloadContext";
-import { TabContext } from "./TabContext";
 
 import "@rsc-parser/core/style.css";
 
@@ -72,323 +50,15 @@ export function Parser() {
         />
       </form>
       <div className="min-h-[calc(100vh-120px)] w-full max-w-full">
-        <PayloadContext.Provider value={payload}>
-          <ErrorBoundary
-            FallbackComponent={GenericErrorBoundaryFallback}
-            key={payload}
-          >
-            <Tabs payload={payload} />
-          </ErrorBoundary>
-        </PayloadContext.Provider>
-      </div>
-    </div>
-  );
-}
-
-function Tabs({ payload }: { payload: string }) {
-  const [isPending, startTransition] = useTransition();
-  const [selectedTab, setSelectedTab] = useState<string | null | undefined>(
-    null
-  );
-  const [currentTab, setCurrentTab] = useState<string | null | undefined>(null);
-
-  const payloadSize = parseFloat(stringToKiloBytes(payload));
-  const rows = splitToCleanRows(payload);
-
-  const selectTab = (nextTab: string | null | undefined) => {
-    if (nextTab !== selectedTab) {
-      setSelectedTab(nextTab);
-      startTransition(() => {
-        setCurrentTab(nextTab);
-      });
-    }
-  };
-
-  const tab = Ariakit.useTabStore({
-    selectedId: selectedTab,
-    setSelectedId: selectTab,
-  });
-
-  return (
-    <TabContext.Provider value={tab}>
-      <div className="flex w-full flex-col items-center justify-center gap-2 px-2 py-6 md:max-w-7xl">
-        <Ariakit.TabList
-          store={tab}
-          className="flex max-w-full flex-row gap-2 overflow-x-auto rounded-2xl !p-2 outline outline-2 outline-offset-2 outline-transparent transition-all duration-200 focus:outline-blue-400 md:flex-wrap md:pb-0"
-        >
-          {rows.map((row) => (
-            <Ariakit.Tab
-              className="group rounded-xl border-none text-left outline outline-2 outline-offset-2 outline-transparent transition-all duration-200 focus:outline-blue-400"
-              key={row}
-              id={row}
-            >
-              <ErrorBoundary
-                fallbackRender={({ error }) => (
-                  <TabFallback
-                    error={error}
-                    row={row}
-                    payloadSize={payloadSize}
-                  />
-                )}
-              >
-                <TabContent row={row} payloadSize={payloadSize} />
-              </ErrorBoundary>
-            </Ariakit.Tab>
-          ))}
-        </Ariakit.TabList>
-
-        <div>Total size: {stringToKiloBytes(payload)} KB (uncompressed)</div>
-      </div>
-
-      <Ariakit.TabPanel
-        store={tab}
-        tabId={currentTab}
-        alwaysVisible={true}
-        className="w-screen max-w-7xl rounded-3xl bg-slate-100 p-4 outline outline-2 outline-offset-2 outline-transparent transition-all delay-100 duration-200 focus:outline-blue-400 dark:bg-slate-800 md:px-12"
-        aria-label="Lines"
-        aria-busy={isPending}
-        style={{
-          opacity: isPending ? "0.6" : "1",
-        }}
-      >
-        {payload === "" ? (
-          <p>Please enter a payload to see results</p>
-        ) : selectedTab === null || selectTab === undefined ? (
-          <p>Please select a tab</p>
-        ) : null}
-
-        {rows
-          .filter((row) => row == currentTab)
-          .map((row) => (
-            <ErrorBoundary
-              FallbackComponent={GenericErrorBoundaryFallback}
-              key={`tab-panel-${row}`}
-            >
-              <TabPanelContent row={row} payloadSize={payloadSize} />
-            </ErrorBoundary>
-          ))}
-      </Ariakit.TabPanel>
-    </TabContext.Provider>
-  );
-}
-
-function TabFallback({
-  error,
-  row,
-  payloadSize,
-}: {
-  error: Error;
-  row: string;
-  payloadSize: number;
-}) {
-  const rowSize = parseFloat(stringToKiloBytes(row));
-
-  if (error instanceof Error) {
-    return (
-      <div className="flex h-full flex-col rounded-xl bg-red-200 px-2 py-1 transition-all duration-200 group-aria-selected:bg-red-600 group-aria-selected:text-white">
-        <div>Error</div>
-        <Meter fraction={rowSize / payloadSize} />
-      </div>
-    );
-  }
-
-  return <span>Error</span>;
-}
-
-function TabContent({
-  row,
-  payloadSize,
-}: {
-  row: string;
-  payloadSize: number;
-}) {
-  const rowSize = parseFloat(stringToKiloBytes(row));
-  const tokens = lexer(row);
-  const { identifier, type } = parse(tokens);
-  const refinedType = refineRowType(type);
-
-  return (
-    <div className="flex flex-row gap-1.5 rounded-xl bg-slate-200 px-2 py-1 transition-all duration-150 group-aria-selected:bg-blue-300  dark:bg-slate-800  dark:group-aria-selected:bg-blue-700">
-      <div className="-mt-px text-xl font-semibold">{identifier}</div>
-      <div className="flex flex-col items-start">
-        <div className="whitespace-nowrap">{refinedType}</div>
-        <Meter fraction={rowSize / payloadSize} />
-      </div>
-    </div>
-  );
-}
-
-function TabPanelContent({
-  row,
-  payloadSize,
-}: {
-  row: string;
-  payloadSize: number;
-}) {
-  return (
-    <div className="flex flex-col gap-6">
-      <div className="flex flex-row justify-between">
         <ErrorBoundary
           FallbackComponent={GenericErrorBoundaryFallback}
-          key={`meta-${row.toString()}`}
+          key={payload}
         >
-          <TabPanelMeta row={row} />
-        </ErrorBoundary>
-        <ErrorBoundary
-          FallbackComponent={GenericErrorBoundaryFallback}
-          key={`size-${row.toString()}`}
-        >
-          <TabPanelSize row={row} payloadSize={payloadSize} />
+          <div className="flex w-full flex-col items-center justify-center gap-2 px-2 py-6 md:max-w-7xl">
+            <RowTabs payload={payload} />
+          </div>
         </ErrorBoundary>
       </div>
-
-      <div className="h-0.5 w-full bg-slate-300 dark:bg-slate-600" />
-
-      <ErrorBoundary
-        FallbackComponent={GenericErrorBoundaryFallback}
-        key={`row-${row.toString()}`}
-      >
-        <TabPanelExplorer row={row} />
-      </ErrorBoundary>
-
-      <div className="h-0.5 w-full bg-slate-300 dark:bg-slate-600" />
-
-      <ErrorBoundary
-        FallbackComponent={GenericErrorBoundaryFallback}
-        key={`tree.${row.toString()}`}
-      >
-        <TabPanelGenericData row={row} />
-      </ErrorBoundary>
     </div>
   );
-}
-
-function TabPanelMeta({ row }: { row: string }) {
-  const tokens = lexer(row);
-  const { identifier, type } = parse(tokens);
-  const refinedType = refineRowType(type);
-
-  return (
-    <div className="flex flex-col gap-1">
-      <h3 className="inline-block rounded-full text-xl font-bold">
-        {identifier} <span className="text-slate-400">/ $L{identifier}</span>
-      </h3>
-      <h4 className="font-medium">
-        {refinedType}{" "}
-        <span className="text-slate-400">/ &quot;{type}&quot;</span>{" "}
-      </h4>
-    </div>
-  );
-}
-
-function TabPanelSize({
-  row,
-  payloadSize,
-}: {
-  row: string;
-  payloadSize: number;
-}) {
-  const rowSize = parseFloat(stringToKiloBytes(row));
-
-  return (
-    <div className="text-right">
-      <div className="whitespace-nowrap">{rowSize} KB row (uncompressed)</div>
-      <div>{((rowSize / payloadSize) * 100).toFixed(2)}% of total</div>
-      <Meter fraction={rowSize / payloadSize} />
-    </div>
-  );
-}
-
-function TabPanelExplorer({ row }: { row: string }) {
-  const tokens = lexer(row);
-  const { type, data } = parse(tokens);
-
-  const refinedType = refineRowType(type);
-
-  const tab = useContext(TabContext);
-  if (tab === undefined) {
-    throw new Error("TabContext must be used within a TabContext.Provider");
-  }
-
-  const payload = useContext(PayloadContext);
-  if (payload === undefined) {
-    throw new Error(
-      "PayloadContext must be used within a PayloadContext.Provider"
-    );
-  }
-
-  switch (refinedType) {
-    case "client ref":
-      // This is a bit iffy. Should probably have separate names for these.
-      return <ClientReferenceRow data={data} />;
-    case "hint":
-      return <HintRow data={data} />;
-    case "tree":
-      return (
-        <TreeRow
-          data={data}
-          onClickClientReference={(name) => {
-            const buttonIdentifier = name.replace("$L", "");
-
-            const rows = splitToCleanRows(payload);
-
-            for (const row of rows) {
-              const tokens = lexer(row);
-              const { identifier } = parse(tokens);
-
-              if (buttonIdentifier === identifier) {
-                // TODO: Don't hard-code this
-                window.scrollTo(0, 680);
-                tab.setSelectedId(row);
-              }
-            }
-          }}
-        />
-      );
-    case "unknown":
-      throw new Error(`Unknown row type: ${type}`);
-  }
-}
-
-function TabPanelGenericData({ row }: { row: string }) {
-  const tokens = lexer(row);
-  const { data } = parse(tokens);
-
-  return (
-    <div className="flex flex-col gap-2">
-      <Details summary="JSON Parsed Data">
-        <ErrorBoundary
-          FallbackComponent={GenericErrorBoundaryFallback}
-          key={`raw-json-${data.toString()}`}
-        >
-          <RawJson data={data} />
-        </ErrorBoundary>
-      </Details>
-
-      <Details summary="Raw Data">{data}</Details>
-    </div>
-  );
-}
-
-function Details({
-  summary,
-  children,
-}: {
-  summary: string;
-  children: ReactNode;
-}) {
-  return (
-    <details className="rounded-lg bg-slate-300 p-3 dark:bg-slate-700">
-      <summary className="cursor-pointer">{summary}</summary>
-      <div className="pt-4">
-        <pre className="whitespace-normal break-all text-sm">{children}</pre>
-      </div>
-    </details>
-  );
-}
-
-function RawJson({ data }: { data: string }) {
-  const json = JSON.parse(data);
-
-  return <JSONTree data={json} shouldExpandNodeInitially={() => true} />;
 }
