--- packages/website/app/PayloadContext.tsx
@@ -1,3 +1,0 @@
-import { createContext } from "react";
-
-export const PayloadContext = createContext("");
