--- packages/website/app/TabContext.tsx
@@ -1,6 +1,0 @@
-import * as Ariakit from "@ariakit/react";
-import { createContext } from "react";
-
-export const TabContext = createContext<Ariakit.TabStore | undefined>(
-  undefined,
-);
